+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ Develop
 
 Major Features and Improvements
 -------------------------------
-
+- add ``JohnsonSU`` PDF, the Johnson SU distribution.
 
 Breaking changes
 ------------------